### 2024-09-22 #7265 @editorialbot
[PRE REVIEW]: Pooltool: A Python package for realistic billiards simulation
**Submitting author:** @ekiefl (<a href="http://orcid.org/0000-0002-6473-0921">Evan Kiefl</a>)
**Repository:** https://github.com/ekiefl/pooltool
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.1
