### 2024-09-22 #7265 @danielskatz
👋 @ekiefl - thanks for your submission. From the title, I was wondering if this was [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS, but I think you make a solid case in the paper.

So I will next work on finding an editor.

I see you don't have an acknowledgments section, so I assume you don't have any funding or other activities you want to acknowledge?

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
