### 2024-09-22 #7265 @danielskatz
Actually, looking at our current editor load, I'll assign myself to be the editor for this one.

Please suggest ~5 potential reviewers. You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. Or people who aren't in the JOSS system would also be ok.  If you know their GitHub usernames, then please mention them here in this thread (without tagging them with an @).
