### 2024-09-23 #7265 @ekiefl
Hi @danielskatz, it's very nice to meet you. Thanks for getting things up and running so quickly!

> 👋 @ekiefl - thanks for your submission. From the title, I was wondering if this was [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS, but I think you make a solid case in the paper.

🎊

> Actually, looking at our current editor load, I'll assign myself to be the editor for this one.

Nice, thanks for volunteering your time 🙏

> I see you don't have an acknowledgments section, so I assume you don't have any funding or other activities you want to acknowledge?

That's right.

> In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command @editorialbot check references to check again, and the command @editorialbot generate pdf when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Sounds good, I'll work on this.

> @ekiefl if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @)

My understanding was that projects accepted in pyOpenSci are processed differently during the JOSS review, although the details of how is what I'm not clear about. Given this, I just wanted to double check that you are aware that pooltool has been accepted in pyOpenSci (submission/review here: https://github.com/pyOpenSci/software-submission/issues/173). I'll wait to hear from you before I start looking for yet more reviewers.
