### 2024-09-23 #7265 @danielskatz
Thanks for mentioning the pyOpenSci part - I missed that.  Yes, then we don't need new reviewers and this will happen much more quickly...
