### 2024-09-25 #7265 @danielskatz
Please do ping me when the DOIs are fixed, then we can move this along.
