### 2024-09-28 #7265 @ekiefl
Hey @danielskatz, I've updated the references with DOIs and they are looking good in the PDF. Thanks for your patience, and please let me know of any next steps.
