### 2024-09-28 #7301 @danielskatz
@ekiefl - please add your country to your affiliation.

Also, in the references:
* the first reference seems to have an incomplete venue/publisher
* Bhagat is not a book but a thesis, so the bibtex should be updated to match https://www.proquest.com/openview/8f032427778440f05e304384c698ec78/1
* For Greenspan et al., please remove the {}s around Pool-Playing
* Similarly for other entries with un-needed {}s, such as Leckie & Greenspan, Mathavan et al.
* Is Sang also a thesis? If so, please update the bibtex
