### 2024-09-28 #7301 @ekiefl
@danielskatz, I've addressed these changes in https://github.com/ekiefl/pooltool/pull/149, which is now merged.

> * Is Sang also a thesis? If so, please update the bibtex

Yes, I've updated accordingly.
