### 2024-09-28 #7301 @danielskatz
Note: expedited review as this was already reviewed for pyOpenSci: https://github.com/pyOpenSci/software-submission/issues/173
