### 2024-09-28 #7301 @danielskatz
@ekiefl - I'll go ahead and accept this, but wanted to point out a few things about the Zenodo archive:
1. The license is listed as CC-BY, rather than Apache-2, which I think it should be
2. Normally, JOSS would request that you change the metadata for the deposit in Zenodo (which you can do manually without creating a new deposit) so that the title matches the title of the paper, but given the fact that this was already accepted for pyOpenSci, this is not required.
