### 2024-09-28 #7301 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
