### 2024-09-28 #7301 @danielskatz
👋 @ekiefl (Evan Kiefl) - Congratulations on your JOSS paper publication (associated with the previous pyOpenSci acceptance of your package)!
